+++ segment/src/convert-str-slice.rs
@@ -1,4 +1,4 @@
-/// Converts a 3-digit string into a Vector of its individual digits (u8).
+/// Converts a 3-digit string into a Vector of its individual digits (u32).
 ///
 /// Example: "132" -> [1, 3, 2]
 ///
@@ -10,18 +10,18 @@
 /// * `input_str` - A reference to the 3-digit string.
 ///
 /// # Returns
-/// A `Vec<u8>` containing the three digits.
-fn string_to_digit_vec(input_str: &str) -> Vec<u8> {
+/// A `Vec<u32>` containing the three digits.
+fn string_to_digit_vec(input_str: &str) -> Vec<u32> {
     input_str
         // Iterate over the characters of the string
         .chars()
         // Convert each character digit (e.g., '1') to a numeric digit (e.g., 1)
         .map(|c| {
             // c.to_digit(10) returns an Option<u32>. Since we are guaranteed 
-            // valid digits, we can safely unwrap and cast to u8.
-            c.to_digit(10).expect("Input must contain only valid digits 0-9") as u8
+            // valid digits, we can safely unwrap and cast to u32.
+            c.to_digit(10).expect("Input must contain only valid digits 0-9") as u32
         })
-        // Collect the results into a new Vector of u8
+        // Collect the results into a new Vector of u32
         .collect()
 }
